update current unit tests with comments change stub from swift to json
